requirements.txt
google-adk>=1.5.0,<2.0.0
google-genai>=1.9.0
google-cloud-aiplatform[agent_engines]>=1.133.0
python-dotenv==1.0.0
pydantic
mcp>=1.0.0
fastmcp
